+++ src/main/java/com/eacuamba/dev/reddit_clone_using_angular_spring/api/data_transfer_object/request/UsernameAndPasswordAuthenticationRequest.java
@@ -1,4 +1,4 @@
-package com.eacuamba.dev.reddit_clone_using_angular_spring.api.data_transfer_object;
+package com.eacuamba.dev.reddit_clone_using_angular_spring.api.data_transfer_object.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
